+++ lib/Remote/RemoteWebDriver.php
@@ -3,8 +3,10 @@
 namespace Facebook\WebDriver\Remote;
 
 use Facebook\WebDriver\Exception\UnsupportedOperationException;
+use Facebook\WebDriver\Html5\WebStorageInterface;
 use Facebook\WebDriver\Interactions\WebDriverActions;
 use Facebook\WebDriver\JavaScriptExecutor;
+use Facebook\WebDriver\Remote\Html5\RemoteWebStorage;
 use Facebook\WebDriver\WebDriver;
 use Facebook\WebDriver\WebDriverBy;
 use Facebook\WebDriver\WebDriverCapabilities;
@@ -469,6 +471,16 @@ class RemoteWebDriver implements WebDriver, JavaScriptExecutor, WebDriverHasInpu
     public function switchTo()
     {
         return new RemoteTargetLocator($this->getExecuteMethod(), $this, $this->isW3cCompliant);
+    }
+
+    /**
+     * An abstraction allowing the driver to access the browser's storage.
+     *
+     * @return WebStorageInterface
+     */
+    public function storage()
+    {
+        return new RemoteWebStorage($this->getExecuteMethod());
     }
 
     /**